Everyone user added. closes #18

+++ app/rest/tests.py
@@ -27,11 +27,9 @@ class SimpleTest(TestCase):
         users = User.retrieve_all()
         for user in users:
             # create the user
-            user.delete()
-            time.sleep(0.1)
-                
-        
-
+            if user.username != 'everyone':
+                user.delete()
+                time.sleep(0.1)
 
     # create repositories
     def create_repos(self):
@@ -83,7 +81,7 @@ class SimpleTest(TestCase):
         self.assertEqual(response.status_code, 200)
         time.sleep(0.1)
         response = self.c.get('/rest/user/')
-        self.assertEqual(response.content, '["user1", "user2"]')
+        self.assertEqual(response.content, '["user1", "user2", "everyone"]')
         
     # update a user pas
     def test_user_change_password(self):
@@ -95,13 +93,13 @@ class SimpleTest(TestCase):
         self.assertEqual(self.c.post('/rest/user/', { 'username' : 'user4', 'password' : 'user4' }).status_code, 200)
         time.sleep(0.1)
         response = self.c.get('/rest/user/')
-        self.assertEqual(response.content, '["user1", "user2", "user3", "user4"]')
+        self.assertEqual(response.content, '["user1", "user2", "user3", "user4", "everyone"]')
         
     # retrieve
     def test_user_retrieve(self):
         response = self.c.get('/rest/user/')
         self.assertEqual(response.status_code, 200)
-        self.assertEqual(response.content, '["user1", "user2", "user3"]')
+        self.assertEqual(response.content, '["user1", "user2", "user3", "everyone"]')
         
     #########################
     # Repository user management
